+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Automatically default packager host to build machine when running a dev build on device - alloy
 * Handle errors that can occur when using the camera and move UIImagePickerController code into Emission - alloy
 * [Dev] Moved to CocoaPods keys - orta
+* [Dev] Make it possible to impersonate a user given their ID and an access token - alloy
 
 ### 1.4.1